+++ R/method_weighting_class.R
@@ -10,30 +10,11 @@
 #'
 #' @keywords internal
 #' @include method_class.R
-#' @include bootstrap_class.R
 .method_weighting_obj <- setClass(
   "method_weighting_obj",
   contains = "method_primary_obj",
-  slots = c(
-    optimal_weight_flag = "logical",
-    wt = "numeric",
-    model_form_piS = "character",
-    model_form_piA = "character",
-    model_form_mu0_ext = "character",
-    model_form_mu0_rct = "character",
-    model_form_mu1_rct = "character"
-  ),
   prototype = list(
-    method_name = "IPW",
-    optimal_weight_flag = FALSE,
-    wt = 0,
-    bootstrap_flag = FALSE,
-    bootstrap_obj = .bootstrap_obj(),
-    model_form_piA = "",
-    model_form_piS = "",
-    model_form_mu0_ext = "",
-    model_form_mu0_rct = "",
-    model_form_mu1_rct = ""
+    method_name = "IPW"
   )
 )